requirements.txt
flask
numpy
Pillow
opencv-python
scikit-learn
